backend/functions/upload-service/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
Pillow==10.0.1
